Created repo

+++ src/main/java/com/tolrom/springlibrary/exception/BookNotFoundException.java
@@ -1,7 +1,7 @@
 package com.tolrom.springlibrary.exception;
 
 public class BookNotFoundException extends RuntimeException {
-    public BookNotFoundException(Integer id) {
+    public BookNotFoundException(Long id) {
         super("The book with id : " + id + " doesn't exist");
     }
 }
